homectl enable: Create links with consistent names
If we create links using mktemp, then "homectl disable" followed
"homectl enable" on a link that is checked in to Git will cause Git to
report there is a difference, even if there isn't one (morally
speaking).  Fixing this up manually is annoying and potentially
confusing.

Instead, sanitize the link target and use it for the link name as well.
That way we are guaranteed to use the same link name for the same link
target every time, and a disable/enable pair won't cause Git to report
spurious changes.