homectl refresh: Use the right package path
As part of commit e11b589a, wherein "homectl list" was changed to report
full paths.  However, I forgot to update "homectl refresh", which was
still expecting "homectl list" to return only the name of each package
and not the path.  As a result, we were passing bogus paths to
enable-internal.

Don't prepend the path when calling enable-internal from refresh, since
it's there already.