hc: A rewrite of homectl in Python
This is the beginning of what will become homectl 0.3 -- which actually
maintains a tree of symlinks similar to homebrew (so as to avoid making
a mess of $PATH etc.).

It also supports OS and arch-dependent components inside a package, so
if a homectl package is a binary package, only relevant binaries will be
made available in $PATH etc.