loader-emacs: Teach `homectl-package` about `package-selected-packages`
There is a bug in `package-install` where `package-selected-packages` is
not updated if it's empty.  We explicitly update it after installing a
package, to ensure the user has a complete set of selected packages and
`package` won't inadvertently uninstall something installed by
`homectl-require`.

We also explicitly save `package-selected-packages` to the user's
custom-file, so that a difference in installed packages will be
noticeable/is recorded correctly in the user's homectl setup (and can be
committed to git along with the change that actually installed the
package).